Delta time is added to a list every frame (no write yet)
