+++ Assets/Scenes/SampleScene.unity
@@ -450,7 +450,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1916968977}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3a25685a56d144665a612f7cbe50f6ef, type: 3}
+  m_Script: {fileID: 11500000, guid: a8f1b43997960436b8690fb2d1391d33, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!4 &1916968979
