+++ Assets/Scripts/FpsLogger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3a25685a56d144665a612f7cbe50f6ef
+guid: a8f1b43997960436b8690fb2d1391d33
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
