--- Assets/Scripts/FpsScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3a25685a56d144665a612f7cbe50f6ef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
