--- Assets/Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 154c0e944e56940d384511eb98d49fb3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
